chore: move send_refresh out of github_events
This will avoid most circular import isssues
