+++ mergify_engine/actions/refresh.py
@@ -16,7 +16,6 @@ import typing
 from mergify_engine import actions
 from mergify_engine import check_api
 from mergify_engine import context
-from mergify_engine import github_events
 from mergify_engine import rules
 from mergify_engine import utils
 
@@ -30,7 +29,7 @@ class RefreshAction(actions.Action):
         self, ctxt: context.Context, rule: rules.EvaluatedRule
     ) -> check_api.Result:
         with utils.aredis_for_stream() as redis_stream:
-            await github_events.send_refresh(
+            await utils.send_refresh(
                 ctxt.redis,
                 redis_stream,
                 ctxt.pull["base"]["repo"],
